Fixes 'estimate rewareds' + fixe parseint to parsefloat

+++ src/utils/cryptoUnitConverter.ts
@@ -1,7 +1,7 @@
 import web3 from "web3";
 
 export function fullOrbsFromWeiOrbs(weiOrbsString: string): number {
-  return parseInt(web3.utils.fromWei(weiOrbsString, "ether"));
+  return parseFloat(web3.utils.fromWei(weiOrbsString, "ether"));
 }
 
 export function weiOrbsFromFullOrbs(fullOrbs: number): bigint {
